+++ CHANGES.md
@@ -1,11 +1,13 @@
 <!--
- - SPDX-FileCopyrightText: 2020 Serokell <https://serokell.io>
+ - SPDX-FileCopyrightText: 2021 Serokell <https://serokell.io>
  -
  - SPDX-License-Identifier: MPL-2.0
  -->
 
 Unreleased
 ==========
+* [#66](https://github.com/serokell/xrefcheck/pull/66)
+  + Added support for ftp links.
 
 0.2
 ==========